Address governing segment review feedback
